adding the csrf component with the logic so we can get the csrf token from the backend to use in our forms and pass the csrf protectiong that django has
